requirements.txt
python-telegram-bot~=22.0
dotenv~=0.9.9
python-dotenv~=1.1.0
langchain~=0.3.25